test/inspector/debugger/step-into-break-on-async-call-expected.txt

Test for Debugger.stepInto with breakOnAsyncCall.

Running test: testSetTimeout
(anonymous) (test.js:0:0)
scheduledAsyncTaskId is set
setTimeout (test.js:0:11)
scheduledAsyncTaskId is empty

Running test: testPromiseThen
(anonymous) (test.js:0:2)
scheduledAsyncTaskId is set
p.then (test.js:0:7)
scheduledAsyncTaskId is empty
